move LoginServerConfig to login package, rename to MessagingConfig
